// src/gen/netmap/service_pb.ts
export interface LocalNodeInfoRequest {
  Body?: LocalNodeInfoRequest_Body;
  MetaHeader?: NeoFsV2Session.RequestMetaHeader;
  VerifyHeader?: NeoFsV2Session.RequestVerificationHeader;
}

export class LocalNodeInfoRequestImpl implements LocalNodeInfoRequest {
  Body?: LocalNodeInfoRequest_Body;
  MetaHeader?: NeoFsV2Session.RequestMetaHeader;
  VerifyHeader?: NeoFsV2Session.RequestVerificationHeader;

  constructor(data?: Partial<LocalNodeInfoRequest>) {
    this.Body = data?.Body ?? undefined;
    this.MetaHeader = data?.MetaHeader ?? undefined;
    this.VerifyHeader = data?.VerifyHeader ?? undefined;
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    if (this.Body) {
      writer.writeMessage(1, this.Body);
    }
    if (this.MetaHeader) {
      writer.writeMessage(2, this.MetaHeader);
    }
    if (this.VerifyHeader) {
      writer.writeMessage(3, this.VerifyHeader);
    }
    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): LocalNodeInfoRequestImpl {
    const reader = new BinaryReader(data);
    const message = new LocalNodeInfoRequestImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        case 1: // Body
          if (wireType === 2) { // Length-delimited
            message.Body = reader.readMessage(LocalNodeInfoRequest_BodyImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        case 2: // MetaHeader
          if (wireType === 2) { // Length-delimited
            message.MetaHeader = reader.readMessage(NeoFsV2Session.RequestMetaHeaderImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        case 3: // VerifyHeader
          if (wireType === 2) { // Length-delimited
            message.VerifyHeader = reader.readMessage(NeoFsV2Session.RequestVerificationHeaderImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): LocalNodeInfoRequestData {
    return {
      Body: this.Body,
      MetaHeader: this.MetaHeader,
      VerifyHeader: this.VerifyHeader
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class LocalNodeInfoRequest extends LocalNodeInfoRequestImpl {}

// Type for interface data only (excludes class methods from merged type)
export type LocalNodeInfoRequestData = Omit<LocalNodeInfoRequest, 'serializeBinary' | 'toObject' | 'constructor'>;

export interface LocalNodeInfoRequest_Body {
}

export class LocalNodeInfoRequest_BodyImpl implements LocalNodeInfoRequest_Body {

  constructor(data?: Partial<LocalNodeInfoRequest_Body>) {
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): LocalNodeInfoRequest_BodyImpl {
    const reader = new BinaryReader(data);
    const message = new LocalNodeInfoRequest_BodyImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): LocalNodeInfoRequest_BodyData {
    return {
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class LocalNodeInfoRequest_Body extends LocalNodeInfoRequest_BodyImpl {}

// Type for interface data only (excludes class methods from merged type)
export type LocalNodeInfoRequest_BodyData = Omit<LocalNodeInfoRequest_Body, 'serializeBinary' | 'toObject' | 'constructor'>;
// ...
export interface NetworkInfoRequest {
  Body?: NetworkInfoRequest_Body;
  MetaHeader?: NeoFsV2Session.RequestMetaHeader;
  VerifyHeader?: NeoFsV2Session.RequestVerificationHeader;
}

export class NetworkInfoRequestImpl implements NetworkInfoRequest {
  Body?: NetworkInfoRequest_Body;
  MetaHeader?: NeoFsV2Session.RequestMetaHeader;
  VerifyHeader?: NeoFsV2Session.RequestVerificationHeader;

  constructor(data?: Partial<NetworkInfoRequest>) {
    this.Body = data?.Body ?? undefined;
    this.MetaHeader = data?.MetaHeader ?? undefined;
    this.VerifyHeader = data?.VerifyHeader ?? undefined;
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    if (this.Body) {
      writer.writeMessage(1, this.Body);
    }
    if (this.MetaHeader) {
      writer.writeMessage(2, this.MetaHeader);
    }
    if (this.VerifyHeader) {
      writer.writeMessage(3, this.VerifyHeader);
    }
    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): NetworkInfoRequestImpl {
    const reader = new BinaryReader(data);
    const message = new NetworkInfoRequestImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        case 1: // Body
          if (wireType === 2) { // Length-delimited
            message.Body = reader.readMessage(NetworkInfoRequest_BodyImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        case 2: // MetaHeader
          if (wireType === 2) { // Length-delimited
            message.MetaHeader = reader.readMessage(NeoFsV2Session.RequestMetaHeaderImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        case 3: // VerifyHeader
          if (wireType === 2) { // Length-delimited
            message.VerifyHeader = reader.readMessage(NeoFsV2Session.RequestVerificationHeaderImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): NetworkInfoRequestData {
    return {
      Body: this.Body,
      MetaHeader: this.MetaHeader,
      VerifyHeader: this.VerifyHeader
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class NetworkInfoRequest extends NetworkInfoRequestImpl {}

// Type for interface data only (excludes class methods from merged type)
export type NetworkInfoRequestData = Omit<NetworkInfoRequest, 'serializeBinary' | 'toObject' | 'constructor'>;

export interface NetworkInfoRequest_Body {
}

export class NetworkInfoRequest_BodyImpl implements NetworkInfoRequest_Body {

  constructor(data?: Partial<NetworkInfoRequest_Body>) {
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): NetworkInfoRequest_BodyImpl {
    const reader = new BinaryReader(data);
    const message = new NetworkInfoRequest_BodyImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): NetworkInfoRequest_BodyData {
    return {
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class NetworkInfoRequest_Body extends NetworkInfoRequest_BodyImpl {}

// Type for interface data only (excludes class methods from merged type)
export type NetworkInfoRequest_BodyData = Omit<NetworkInfoRequest_Body, 'serializeBinary' | 'toObject' | 'constructor'>;
// ...
export interface NetmapSnapshotRequest {
  Body?: NetmapSnapshotRequest_Body;
  MetaHeader?: NeoFsV2Session.RequestMetaHeader;
  VerifyHeader?: NeoFsV2Session.RequestVerificationHeader;
}

export class NetmapSnapshotRequestImpl implements NetmapSnapshotRequest {
  Body?: NetmapSnapshotRequest_Body;
  MetaHeader?: NeoFsV2Session.RequestMetaHeader;
  VerifyHeader?: NeoFsV2Session.RequestVerificationHeader;

  constructor(data?: Partial<NetmapSnapshotRequest>) {
    this.Body = data?.Body ?? undefined;
    this.MetaHeader = data?.MetaHeader ?? undefined;
    this.VerifyHeader = data?.VerifyHeader ?? undefined;
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    if (this.Body) {
      writer.writeMessage(1, this.Body);
    }
    if (this.MetaHeader) {
      writer.writeMessage(2, this.MetaHeader);
    }
    if (this.VerifyHeader) {
      writer.writeMessage(3, this.VerifyHeader);
    }
    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): NetmapSnapshotRequestImpl {
    const reader = new BinaryReader(data);
    const message = new NetmapSnapshotRequestImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        case 1: // Body
          if (wireType === 2) { // Length-delimited
            message.Body = reader.readMessage(NetmapSnapshotRequest_BodyImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        case 2: // MetaHeader
          if (wireType === 2) { // Length-delimited
            message.MetaHeader = reader.readMessage(NeoFsV2Session.RequestMetaHeaderImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        case 3: // VerifyHeader
          if (wireType === 2) { // Length-delimited
            message.VerifyHeader = reader.readMessage(NeoFsV2Session.RequestVerificationHeaderImpl.deserializeBinary);
          } else {
            reader.skipField(wireType);
          }
          break;
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): NetmapSnapshotRequestData {
    return {
      Body: this.Body,
      MetaHeader: this.MetaHeader,
      VerifyHeader: this.VerifyHeader
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class NetmapSnapshotRequest extends NetmapSnapshotRequestImpl {}

// Type for interface data only (excludes class methods from merged type)
export type NetmapSnapshotRequestData = Omit<NetmapSnapshotRequest, 'serializeBinary' | 'toObject' | 'constructor'>;

export interface NetmapSnapshotRequest_Body {
}

export class NetmapSnapshotRequest_BodyImpl implements NetmapSnapshotRequest_Body {

  constructor(data?: Partial<NetmapSnapshotRequest_Body>) {
  }

  serializeBinary(): Uint8Array {
    const writer = new BinaryWriter();

    return writer.getResultBuffer();
  }

  static deserializeBinary(data: Uint8Array): NetmapSnapshotRequest_BodyImpl {
    const reader = new BinaryReader(data);
    const message = new NetmapSnapshotRequest_BodyImpl();

    // Parse protobuf wire format
    while (reader.position < reader.buffer.length) {
      const tag = reader.readVarint();
      const fieldNumber = tag >>> 3;
      const wireType = tag & 7;

      switch (fieldNumber) {
        default:
          // Skip unknown fields
          reader.skipField(wireType);
          break;
      }
    }

    return message;
  }

  toObject(): NetmapSnapshotRequest_BodyData {
    return {
    };
  }

}

// Export class alias without Impl suffix for Node.js compatibility
export class NetmapSnapshotRequest_Body extends NetmapSnapshotRequest_BodyImpl {}

// Type for interface data only (excludes class methods from merged type)
export type NetmapSnapshotRequest_BodyData = Omit<NetmapSnapshotRequest_Body, 'serializeBinary' | 'toObject' | 'constructor'>;